+++ ChangeLog.md
@@ -1,8 +1,16 @@
 # ChangeLog
 KsWare.Presentation.ViewFramework.Common
 
-#### 2019-12-24 (v0.1.x)
-- Multiframework Support (.Net 4.5+, Core 3.x)
+### 2020-01-25
+- RootBindingExtension
+- RootElementExtension
+- SharedWidth
+- DesignTime.Properties
+
 
 #### 2019-09-03 (v0.1.32)
 - VisibilityBinding
+- 
+#### 2019-12-24 (v0.1.x)
+- Multiframework Support (.Net 4.5+, Core 3.x)
+